docs(observability): updates observability docs with up-to-date Greenhouse architecture
This is a documentation update is the current setup as we in the Greenhouse Core (and Observability Team) has designed it. We would urge that the ApeiroRA documentation would mirror this somehow such that we can be consistent in our communication.

In particular it
- adds more context to the signals, tools and applications that Greenhouse is providing through its suite of Plugins.
- adds two updated architecture diagrams show-casing a high-level and a detailed view into how these Plugins can operate with one and other.
- adds term definitions for Greenhouse and Greenhouse Plugins.

For further information about the technical discussion making this page and diagrams see here https://github.com/cloudoperators/greenhouse/pull/1154

---------

on-behalf-of: @SAP Simon Olander (simon.olander@sap.com)